Revert "refactor : 기존 코드 구조를 mvc패턴으로 구조 리팩토링"
This reverts commit 42ef1c1936d6c869817fdae16fd172481763ef9a.
