--- src/main/java/calculator/exception/ErrorMessage.java
@@ -1,8 +1,0 @@
-package calculator.exception;
-
-public class ErrorMessage {
-    public static final String NEGATIVE_NUMBER = "음수는 입력할 수 없습니다: ";
-    public static final String INVALID_CUSTOM_DELIMITER = "커스텀 구분자 지정이 잘못되었습니다. '//구분자\\n숫자' 형식으로 입력해야 합니다.";
-    public static final String NON_NUMERIC_VALUE = "숫자가 아닌 값이 포함되어 있습니다: ";
-    public static final String INPUT_MUST_END_WITH_NUMBER = "입력은 반드시 숫자로 끝나야합니다.";
-}
