--- src/main/java/calculator/view/InputView.java
@@ -1,10 +1,0 @@
-package calculator.view;
-
-import camp.nextstep.edu.missionutils.Console;
-
-public class InputView {
-
-    public static String readInput() {
-        return Console.readLine();
-    }
-}
